style(*): rename `RequestException*` to `NetworkException*`
