+++ src/main/java/sleppynavigators/studyupbackend/application/authentication/AuthService.java
@@ -22,7 +22,7 @@ import sleppynavigators.studyupbackend.infrastructure.user.UserRepository;
 import sleppynavigators.studyupbackend.presentation.authentication.dto.RefreshRequest;
 import sleppynavigators.studyupbackend.presentation.authentication.dto.SignInRequest;
 import sleppynavigators.studyupbackend.presentation.authentication.dto.TokenResponse;
-import sleppynavigators.studyupbackend.exception.request.InvalidCredentialException;
+import sleppynavigators.studyupbackend.exception.network.InvalidCredentialException;
 
 @Service
 @Transactional(readOnly = true)
