+++ src/main/java/sleppynavigators/studyupbackend/exception/network/ForbiddenException.java
@@ -1,8 +1,8 @@
-package sleppynavigators.studyupbackend.exception.request;
+package sleppynavigators.studyupbackend.exception.network;
 
 import sleppynavigators.studyupbackend.exception.ErrorCode;
 
-public class ForbiddenException extends RequestBaseException {
+public class ForbiddenException extends NetworkBaseException {
 
     protected ForbiddenException(int status, String code, String message) {
         super(status, code, message);
