+++ src/main/java/sleppynavigators/studyupbackend/exception/network/InvalidApiException.java
@@ -1,8 +1,8 @@
-package sleppynavigators.studyupbackend.exception.request;
+package sleppynavigators.studyupbackend.exception.network;
 
 import sleppynavigators.studyupbackend.exception.ErrorCode;
 
-public class InvalidApiException extends RequestBaseException {
+public class InvalidApiException extends NetworkBaseException {
 
     protected InvalidApiException(int status, String code, String message) {
         super(status, code, message);
