+++ src/main/java/sleppynavigators/studyupbackend/exception/network/InvalidCredentialException.java
@@ -1,8 +1,8 @@
-package sleppynavigators.studyupbackend.exception.request;
+package sleppynavigators.studyupbackend.exception.network;
 
 import sleppynavigators.studyupbackend.exception.ErrorCode;
 
-public class InvalidCredentialException extends RequestBaseException {
+public class InvalidCredentialException extends NetworkBaseException {
 
     protected InvalidCredentialException(int status, String code, String message) {
         super(status, code, message);
